Refactor to SOLID principles
