+++ Gemfile
@@ -2,4 +2,5 @@ source 'https://rubygems.org'
 
 # gem "rails"
 
+gem 'pry', '~> 0.13.1'
 gem 'rubocop', '>= 1.0', '< 2.0'
